Refactor the types
