+++ iml-graphql-queries/Cargo.toml
@@ -10,5 +10,6 @@ version = "0.2.0"
 chrono = "0.4"
 humantime = "2.0"
 iml-wire-types = {version = "0.4", path = "../iml-wire-types"}
+juniper = {git = "https://github.com/graphql-rust/juniper"}
 serde = {version = "1", features = ["derive"]}
 serde_json = "1"
